Enhance data structures: add Queue, improve safety & perf
Major improvements across the library:
- Added Queue class with Sll backend, capacity, and full API.
- Improved Array/List exception safety and memory management.
- Sll now uses tail pointer for O(1) addLast/back; fixed shuffle leak.
- Stack: added copy/move, initializer list, setCap, and better checks.
- Refactored BubbleSort/MergeSort; added iterator helpers in utils.
- Updated tests and README to reflect new features and fixes.

+++ include/dll.hpp
@@ -90,7 +90,7 @@ namespace mLib {
             if (this != &other) {
                 Node* moveHead = other.head;
                 Node* moveTail = other.tail;
-                int moveSize = other.size;
+                size_t moveSize = other.size;
 
                 other.size = size;
                 other.head = head;
